+++ hack/tests/scorecard-subcommand.sh
@@ -24,6 +24,10 @@ echo $commandoutput | grep "Total Score: 80%"
 # test config file
 commandoutput2="$(operator-sdk scorecard \
   --proxy-image "$DEST_IMAGE" \
-  --config "$CONFIG_PATH")"
-echo $commandoutput2 | grep "Total Score: 80%"
+  --config "$CONFIG_PATH" \
+  --output json)"
+# check basic suite
+echo $commandoutput2 | grep '^.*"error": 0,[[:space:]]"pass": 3,[[:space:]]"partialPass": 0,[[:space:]]"fail": 0,[[:space:]]"totalTests": 3,[[:space:]]"totalScorePercent": 100,.*$'
+# check olm suite
+echo $commandoutput2 | grep '^.*"error": 0,[[:space:]]"pass": 3,[[:space:]]"partialPass": 0,[[:space:]]"fail": 2,[[:space:]]"totalTests": 5,[[:space:]]"totalScorePercent": 61,.*$'
 popd